Porting to Solaris... This compiles, I have yet to do any real testing. Once this is checked in, I will make sure these changes don't break Linux/FreeBSD.
configure.ac:
    Adding socket,nsl,rt libraries
    Adding check for paths.h header file

ErrLogSyslog.c:
    Define LOG_AUTHPRIV and LOG_FTP if they are not defined.

Options.c:
conf.c:
    Adding cast from char to int for all isspace/isdigit args.

readpassphrase.c:
    remove cdefs.h.
    Add check for autoconf paths.h value, include or define _PATH_TTY.
    Adding cast from char to int for all isspace/isdigit args.

bwctl.c:
    Adding cast from char to int for all isspace/isdigit args.
    Change setenv to putenv to set POSIXLY_CORRECT for getopt.
    Check for LOG_PERROR. (Add warning if -r is passed in, and
        LOG_PERROR is not available.)

bwctld.c:
    Add check for LOG_PERROR.

policy.c:
    Remove fts.h - not even used here, this is an owamp leftover...
    Adding cast from char to int for all isspace/isdigit args.

util.c:
    Adding cast from char to int for all isspace/isdigit args.


git-svn-id: http://anonsvn.internet2.edu/svn/I2util/trunk@114 af504c07-53ac-4b46-a7f1-a0422916b9a2

+++ I2util/Options.c
@@ -155,7 +155,7 @@ static	char	*fmt_opt_string(
 	char	*s;
 
 	for (i=1, s = arg; i<n; i++) {
-		while(*s && ! isspace(*s)) s++;
+		while(*s && ! isspace((int)*s)) s++;
 
 		if (! *s) {
 			I2ErrLog(eh, "Arg string invalid: %s", arg);
@@ -163,7 +163,7 @@ static	char	*fmt_opt_string(
 		}
 		*s++ = '\0';
 
-		while(*s && isspace(*s)) s++;
+		while(*s && isspace((int)*s)) s++;
 	}
 	return(arg);
 }
